Moves hooks to api file to simplify things
