+++ web/src/features/feature-flags/api.ts
@@ -1,4 +1,4 @@
-import { type QueryClient } from '@tanstack/react-query';
+import { useQuery, type QueryClient } from '@tanstack/react-query';
 import { QUERY_KEYS, getBasePath, getHeaders } from 'api/helpers';
 import { FeatureFlags } from './types';
 
@@ -23,4 +23,18 @@ export async function getFeatureFlags(): Promise<FeatureFlags> {
     console.log(error);
     return {};
   }
+}
+
+export function useFeatureFlags(): FeatureFlags {
+  return (
+    useQuery<FeatureFlags>([QUERY_KEYS.FEATURE_FLAGS], async () => getFeatureFlags(), {
+      suspense: true,
+    }).data || {}
+  );
+}
+
+export function useFeatureFlag(name: string): boolean {
+  const features = useFeatureFlags();
+
+  return features?.[name] || false;
 }
